Test fixes for stabilizing tests (#19068)

+++ test/powershell/Language/Scripting/NativeExecution/NativeCommandProcessor.Tests.ps1
@@ -260,6 +260,14 @@ Describe "Run native command from a mounted FAT-format VHD" -tags @("Feature", "
         if (-not $IsWindows) {
             return;
         }
+        else {
+            $storageModule = Get-Module -Name 'Storage' -ListAvailable -ErrorAction SilentlyContinue
+
+            if (-not $storageModule) {
+                Write-Verbose -Verbose "Storage module is not available."
+                return;
+            }
+        }
 
         $vhdx = Join-Path -Path $TestDrive -ChildPath ncp.vhdx
 
@@ -283,11 +291,18 @@ Describe "Run native command from a mounted FAT-format VHD" -tags @("Feature", "
         diskpart.exe /s $create_vhdx
         Mount-DiskImage -ImagePath $vhdx > $null
 
-        Copy-Item "$env:WinDir\System32\whoami.exe" T:\whoami.exe
+        Copy-Item "$env:WinDir\System32\whoami.exe" "T:\whoami.exe"
     }
 
     AfterAll {
         if ($IsWindows) {
+            $storageModule = Get-Module -Name 'Storage' -ListAvailable -ErrorAction SilentlyContinue
+
+            if (-not $storageModule) {
+                Write-Verbose -Verbose "Storage module is not available."
+                return;
+            }
+
             Dismount-DiskImage -ImagePath $vhdx
             Remove-Item $vhdx, $create_vhdx -Force
         }
